[Rust] Move `QualifiedName` to own module
This is used in more places than types, so its best we keep it separate.
